+++ src/test/java/org/stellar/sdk/requests/OffersRequestBuilderTest.java
@@ -43,21 +43,6 @@ public class OffersRequestBuilderTest {
           + "}";
 
   @Test
-  public void testForAccount() {
-    Server server = new Server("https://horizon-testnet.stellar.org");
-    HttpUrl uri =
-        server
-            .offers()
-            .forAccount("GBRPYHIL2CI3FNQ4BXLFMNDLFJUNPU2HY3ZMFSHONUCEOASW7QC7OX2H")
-            .limit(200)
-            .order(RequestBuilder.Order.DESC)
-            .buildUri();
-    assertEquals(
-        "https://horizon-testnet.stellar.org/accounts/GBRPYHIL2CI3FNQ4BXLFMNDLFJUNPU2HY3ZMFSHONUCEOASW7QC7OX2H/offers?limit=200&order=desc",
-        uri.toString());
-  }
-
-  @Test
   public void testWithoutParams() {
     Server server = new Server("https://horizon-testnet.stellar.org");
     HttpUrl uri = server.offers().buildUri();